Move PT_geometry to a new "geometries" folder in the main branch
Parabolic trought geometry of the module studied by R. Dickes
