+++ labothappy/component/solar/parabolic_through/design/parabolic_plant_design.py
@@ -2,7 +2,7 @@ import __init__
 import numpy as np
 import labothappy
 import labothappy.component
-from labothappy.component.solar.parabolic_through.steady_state.modules.parabolic_through_geometry import PT_Collector_Geom
+from labothappy.geometries.solar.parabolic_through_geometry import PT_Collector_Geom
 # from library.component.sizing.solar.parabolic_through.
 
 PT_geom = PT_Collector_Geom()
