+++ labothappy/component/solar/parabolic_through/steady_state/parabolic_through_collector.py
@@ -14,7 +14,7 @@ Rémi Dickes, Vincent Lemort and Sylvain Quoilin
 import parabolic_through_collector
 from CoolProp.CoolProp import PropsSI
 import numpy as np
-from modules.parabolic_through_geometry import PT_Collector_Geom
+from labothappy.geometries.solar.parabolic_through_geometry import PT_Collector_Geom
 from labothappy.component.base_component import BaseComponent
 
 from labothappy.connector.mass_connector import MassConnector
