+++ labothappy/geometries/solar/parabolic_through_geometry.py
